Correxion de nombres Lexer
