+++ lexer/src/main/kotlin/lexer/LexerInterface.kt
@@ -1,0 +1,8 @@
+package lexer
+
+import token.TokenInterface
+
+interface LexerInterface {
+    fun tokenize(input: String): List<TokenInterface>
+    fun iterator(input: String): Iterator<TokenInterface>
+}
